changes in domain classes - nonmutable ID in entities are now nonmutable
git-svn-id: https://acris.googlecode.com/svn@7328 e28c4d36-3818-11de-97fb-aff7d2a6cc54

+++ corpis/branches/gwt_25/corpis-domain/corpis-invoice-domain-model/src/main/java/sk/seges/corpis/server/domain/invoice/Vat.java
@@ -8,7 +8,7 @@ import sk.seges.sesam.domain.IDomainObject;
 
 @DomainInterface
 @BaseObject
-public interface Vat extends IDomainObject<Long> {
+public interface Vat extends IDomainObject<Short> {
 
 	Short vat();
 	Date validFrom();
